AG-10919 - added example that shows adding data to the grid
