+++ documentation/ag-grid-docs/src/content/docs/data-update-row-data/index.mdoc
@@ -119,7 +119,7 @@ For example:
 {% /note %}
 {% /if %}
 
-## Adding new rows
+## Adding New Rows
 
 The example below shows how to add new data to the grid using [Full Row Edit](./cell-editing-full-row/) and [Row Pinning](./row-pinning/)
